API for periodic run
